+++ Assets/Scripts/UI/RadioBinding.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: edabdc631fc80084282d7289e1c821da
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
